修改暂无评论

+++ app/src/main/java/com/secretk/move/bean/FindKwBean.java
@@ -177,7 +177,7 @@ public class FindKwBean extends BaseRes{
                 private String projectCode;
                 private String tokenName;
                 private int tokenCount;
-                private int tokenEveryCount;
+                private double tokenEveryCount;
                 private int tokenCash;
                 private int tokenNum;
                 private int tokenSurplusNum;
@@ -279,11 +279,11 @@ public class FindKwBean extends BaseRes{
                     this.tokenCount = tokenCount;
                 }
 
-                public int getTokenEveryCount() {
+                public double getTokenEveryCount() {
                     return tokenEveryCount;
                 }
 
-                public void setTokenEveryCount(int tokenEveryCount) {
+                public void setTokenEveryCount(double tokenEveryCount) {
                     this.tokenEveryCount = tokenEveryCount;
                 }
